Fix cpp and jvm code of pool arrays
