+++ kt/godot-runtime/src/main/kotlin/godot/core/Variant.kt
@@ -1,6 +1,5 @@
 package godot.core
 
-import godot.core.AABB
 import godot.core.VariantType.*
 import godot.util.toRealT
 import java.nio.ByteBuffer
@@ -18,24 +17,24 @@ val variantMapper = mutableMapOf(
         AABB::class to AABB,
         Basis::class to BASIS,
         Color::class to COLOR,
-//        Dictionary::class to VariantType.DICTIONARY,
-//        GodotArray::class to VariantType.ARRAY,
+        Dictionary::class to DICTIONARY,
+        VariantArray::class to ARRAY,
         Plane::class to PLANE,
-//        NodePath::class to VariantType.NODE_PATH,
+        NodePath::class to NODE_PATH,
         Quat::class to QUAT,
         Rect2::class to RECT2,
-//        RID::class to VariantType.RID,
+        RID::class to _RID,
         Transform::class to TRANSFORM,
         Transform2D::class to TRANSFORM2D,
         Vector2::class to VECTOR2,
         Vector3::class to VECTOR3,
-        PoolByteArray::class to VariantType.POOL_BYTE_ARRAY,
-        PoolColorArray::class to VariantType.POOL_COLOR_ARRAY,
-        PoolIntArray::class to VariantType.POOL_INT_ARRAY,
-        PoolRealArray::class to VariantType.POOL_REAL_ARRAY,
-        PoolStringArray::class to VariantType.POOL_STRING_ARRAY,
-        PoolVector2Array::class to VariantType.POOL_VECTOR2_ARRAY,
-        PoolVector3Array::class to VariantType.POOL_VECTOR3_ARRAY,
+        PoolByteArray::class to POOL_BYTE_ARRAY,
+        PoolColorArray::class to POOL_COLOR_ARRAY,
+        PoolIntArray::class to POOL_INT_ARRAY,
+        PoolRealArray::class to POOL_REAL_ARRAY,
+        PoolStringArray::class to POOL_STRING_ARRAY,
+        PoolVector2Array::class to POOL_VECTOR2_ARRAY,
+        PoolVector3Array::class to POOL_VECTOR3_ARRAY,
 )
 
 var ByteBuffer.bool: Boolean
@@ -329,46 +328,54 @@ enum class VariantType(
 
     // arrays
     POOL_BYTE_ARRAY(
-            { buffer: ByteBuffer, expectedType: Int ->
+            { buffer: ByteBuffer, _: Int ->
                 val ptr = buffer.long
+                PoolByteArray(ptr)
             },
             { buffer: ByteBuffer, any: Any ->
                 POOL_BYTE_ARRAY.toGodotNativeCoreType<PoolByteArray>(buffer, any)
             }
     ), // 20
     POOL_INT_ARRAY(
-            { buffer: ByteBuffer, expectedType: Int ->
+            { buffer: ByteBuffer, _: Int ->
                 val ptr = buffer.long
+                PoolIntArray(ptr)
             },
             { buffer: ByteBuffer, any: Any ->
-                POOL_Int_ARRAY.toGodotNativeCoreType<PoolIntArray>(buffer, any)
+                POOL_INT_ARRAY.toGodotNativeCoreType<PoolIntArray>(buffer, any)
             }
     ),
     POOL_REAL_ARRAY(
-            { buffer: ByteBuffer, expectedType: Int ->
+            { buffer: ByteBuffer, _: Int ->
                 val ptr = buffer.long
+                PoolRealArray(ptr)
             },
             { buffer: ByteBuffer, any: Any ->
                 POOL_REAL_ARRAY.toGodotNativeCoreType<PoolRealArray>(buffer, any)
             }
     ),
     POOL_STRING_ARRAY(
-            { buffer: ByteBuffer, expectedType: Int ->
+            { buffer: ByteBuffer, _: Int ->
                 val ptr = buffer.long
+                PoolStringArray(ptr)
             },
-            { buffer: ByteBuffer, any: Any -> TODO() }
+            { buffer: ByteBuffer, any: Any ->
+                POOL_STRING_ARRAY.toGodotNativeCoreType<PoolStringArray>(buffer, any)
+            }
     ),
     POOL_VECTOR2_ARRAY(
-            { buffer: ByteBuffer, expectedType: Int ->
+            { buffer: ByteBuffer, _: Int ->
                 val ptr = buffer.long
+                PoolVector2Array(ptr)
             },
             { buffer: ByteBuffer, any: Any ->
                 POOL_VECTOR2_ARRAY.toGodotNativeCoreType<PoolVector2Array>(buffer, any)
             }
     ),
     POOL_VECTOR3_ARRAY(
-            { buffer: ByteBuffer, expectedType: Int ->
+            { buffer: ByteBuffer, _: Int ->
                 val ptr = buffer.long
+                PoolVector3Array(ptr)
             },
             { buffer: ByteBuffer, any: Any ->
                 POOL_VECTOR3_ARRAY.toGodotNativeCoreType<PoolVector3Array>(buffer, any)
@@ -377,6 +384,7 @@ enum class VariantType(
     POOL_COLOR_ARRAY(
             { buffer: ByteBuffer, expectedType: Int ->
                 val ptr = buffer.long
+                PoolColorArray(ptr)
             },
             { buffer: ByteBuffer, any: Any ->
                 POOL_COLOR_ARRAY.toGodotNativeCoreType<PoolColorArray>(buffer, any)
